requirements.txt
# Core Flask Framework
flask>=2.0.0,<3.0.0
flask-login>=0.6.0,<1.0.0
flask-bootstrap>=3.3.7,<4.0.0
flask-sqlalchemy>=3.0.0,<4.0.0
flask-socketio>=5.3.0,<6.0.0
flask-limiter>=3.3.0,<4.0.0
flask-migrate>=4.0.4,<5.0.0
flask-wtf>=1.1.1,<2.0.0
flask-mail>=0.9.1,<1.0.0
werkzeug>=2.3.0,<3.0.0

# Security & Authentication
bcrypt>=4.0.1,<5.0.0
python-dotenv>=1.0.0,<2.0.0
email-validator>=2.0.0,<3.0.0
marshmallow>=3.19.0,<4.0.0
flask-jwt-extended>=4.5.2,<5.0.0

# System Monitoring & Security Scanning
psutil>=5.9.0,<6.0.0
scapy>=2.5.0,<3.0.0
pywin32>=305,<400
python-nmap>=0.7.1,<1.0.0
nmap>=0.7.1,<1.0.0  # Added: Direct nmap import found in code

# AI & Machine Learning
torch>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0  # Updated: More specific version
ollama>=0.1.0,<1.0.0
python-ollama>=0.1.0,<1.0.0

# PDF & Report Generation
reportlab>=4.0.0,<5.0.0
fpdf2>=2.7.4,<3.0.0
pytz>=2023.3,<2026.0
beautifulsoup4>=4.12.0,<5.0.0
lxml>=4.9.0,<5.0.0

# Testing Framework
pytest>=7.4.0,<8.0.0
pytest-cov>=4.1.0,<5.0.0
pytest-mock>=3.11.1,<4.0.0
pytest-asyncio>=0.21.1,<1.0.0
pytest-timeout>=2.1.0,<3.0.0
coverage>=7.3.0,<8.0.0

# Caching & Background Tasks
flask_caching>=2.0.0,<3.0.0
redis>=4.6.0,<5.0.0
apscheduler>=3.10.0,<4.0.0

# HTTP & Network
requests>=2.28.0,<3.0.0
aiohttp>=3.8.0,<4.0.0
timeout-decorator>=0.5.0,<1.0.0